build: set coverage threshold to 95% because of c8 bugs
For some reason about <5% of lines c8 marks as uncovered for no apparent
reason. I tried to mark these manually as covered but it seems like
these lines move around. Just set the threshold to 95% as a compromise.
If a large swath of code is uncovered it will probably be more than 5%
anyway.