fix(logo): less thin and blurry
I was rendering the vector image at the default (low) DPI, so it would
come out blurry after resizing it. To fix this I tried to make the text
`\Huge`, which helped a bit but also makes the text thin.

I realized that I can choose a high DPI of 10080 (arbitrary highly
composite number) in the imagemagick `convert` instead.

Additionally, make the text `\tiny` so that it comes out thicker.