doc: explain jargon in README
From @jermiahbuckley: I was going to add this to the discord chat, but
it seems like it might be more useful as a file in the repo. Happy to
reformat this according to any suggested edits, or you can just lift the
contents and reuse in a better way.

@jermiahbuckley: Wrote the whole thing.

yaseen@mowzer.co.za: Wrote commit message. Moved doc into README.md (was
initially in documentation/introduction.md). Fixed Formatting. Made a
few edits.

Co-authored-by: Yaseen Mowzer <yaseen@mowzer.co.za>