feat: add route which returns projectId
For some reason when I deploy to Cloudflare, it erases all environment
variables that are not secret. So I have made all the environment
variable secret. The problem now is that I need the PROJECT_ID
environment variable. But I forget what it was. This is a temporary
route, that will allow me to extract the PROJECT_ID.