feat: make refund bonus and default payment configurable
Add `refundBonusPercent` and `defaultPaymentAmount` to `Project` and in
the App and the worker use these values instead of the hard-coded 20%
and $89.

Since we have projects in production which have the hard-coded amounts
of 20% for the refund bonus percent and $89 for the default payment
amount, I replaced `getProjects` with `getProject` and `setProject`.
`setProject` simply `KV.put`s the project, but `getProject` will fill in
default values of `20` and `89` for `refundBonusPercent` and
`defaultPaymentAmount` respectively.