refactor: validate using zod
There are many places where we serialize JSON objects, and then send
them over the wire, and then parse them on the other side. We are not
validating what we get on the other-side, so it's possible that there
could be weird errors.

Create a module schema.ts that contains zod definitions that can be used
to validate various requests and responses.