feat: add link to application form on project page
Currently there is no way on the website to apply to have your project
hosted. As a simple fix add a link (configurable with the environment
variable REACT_APP_PROJECT_APPLICATION_FORM) on the project page, that
points to a place that one can apply for a project (e.g. a Google Form).